pipeline/requirements.txt
# Data Processing
pandas==2.0.1
numpy==1.24.3
scikit-learn==1.2.2

# Database
sqlalchemy==2.0.12
psycopg2-binary==2.9.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.2
requests==2.30.0

# Logging
loguru==0.7.0

python-json-logger==2.0.7
prometheus-client==0.17.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Security
cryptography==40.0.2

# Data Formats
pyyaml==6.0
json5==0.9.14

# Concurrency
asyncio==3.4.3